Changed CSV parser to immediately normalize row balance values.
This decision simplifies the CSV importer and makes it easier to explain how derived
classes can use it to implement specialized CSV importers. However, it prevents its
results from being cached and shared among accounts that do not share the same sign.
This is alleviated by caching the contents of the CSV file prior to parsing.

Not also that since the identity() method of the CSV importer does not need the
parsed CSV data, the advantage of caching the parsed results are currently nil.